add test

+++ go/src/github.com/w3esoft/glaive/input/input.go
@@ -2,5 +2,5 @@ package input
 
 type Input interface {
 	GetPosition() int
-	Read() int
+	Read() uint8
 }
